Add super admin enpoints

+++ src/features/blogs/api/blogger.controller.ts
@@ -3,6 +3,7 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
   HttpStatus,
@@ -21,6 +22,7 @@ import {
   ApiForbiddenResponse,
   ApiNoContentResponse,
   ApiNotFoundResponse,
+  ApiOkResponse,
   ApiOperation,
   ApiTags,
   ApiUnauthorizedResponse,
@@ -55,7 +57,6 @@ export class BloggerController {
     private readonly commandBus: CommandBus,
   ) {}
 
-  @Post('blogs')
   @ApiOperation({ summary: 'Create new blog' })
   @ApiCreatedResponse({
     status: 201,
@@ -69,6 +70,7 @@ export class BloggerController {
     status: 401,
     description: 'Unauthorized',
   })
+  @Post('blogs')
   async create(
     @Body() createBlogInput: CreateBlogInput,
     @GetCurrentJwtContext() ctx: JwtATPayload,
@@ -81,22 +83,16 @@ export class BloggerController {
     return this.blogsQueryRepository.findOne(blog.id);
   }
 
-  @ApiOperation({ summary: 'Delete post specified by id' })
-  @ApiNoContentResponse({
-    status: 204,
+  @ApiOperation({
+    summary: 'Returns blogs (for which current user is owner) with paging',
+  })
+  @ApiOkResponse({
+    status: 200,
     description: 'No Content',
   })
   @ApiUnauthorizedResponse({
     status: 401,
     description: 'Unauthorized',
-  })
-  @ApiForbiddenResponse({
-    status: 403,
-    description: 'Forbidden',
-  })
-  @ApiNotFoundResponse({
-    status: 404,
-    description: 'Not Found',
   })
   @Get('blogs')
   findAll(
@@ -108,8 +104,6 @@ export class BloggerController {
     );
   }
 
-  @Put('blogs/:id')
-  @HttpCode(HttpStatus.NO_CONTENT)
   @ApiOperation({ summary: 'Update existing Blog by id with InputModel' })
   @ApiNoContentResponse({
     status: 204,
@@ -128,17 +122,25 @@ export class BloggerController {
     description:
       "If user try to update blog that doesn't belong to current user",
   })
-  async update(@Param('id') id: string, @Body() updateBlogDto: UpdateBlogDto) {
+  @Put('blogs/:id')
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async update(
+    @Param('id') id: string,
+    @GetCurrentJwtContext() ctx: JwtATPayload,
+    @Body() updateBlogDto: UpdateBlogDto,
+  ) {
     const blog = await this.blogsService.update(id, updateBlogDto);
     if (!blog) {
       throw new NotFoundException();
     }
 
+    if (blog.userId !== ctx.user.id) {
+      throw new ForbiddenException();
+    }
+
     return;
   }
 
-  @Delete('blogs/:id')
-  @HttpCode(HttpStatus.NO_CONTENT)
   @ApiOperation({ summary: 'Delete blog specified by id' })
   @ApiNoContentResponse({
     status: 204,
@@ -156,18 +158,25 @@ export class BloggerController {
     status: 404,
     description: 'Not Found',
   })
-  async remove(@Param('id') id: string) {
+  @Delete('blogs/:id')
+  @HttpCode(HttpStatus.NO_CONTENT)
+  async remove(
+    @Param('id') id: string,
+    @GetCurrentJwtContext() ctx: JwtATPayload,
+  ) {
     const blog = await this.blogsService.findOne(id);
 
     if (!blog) {
       throw new NotFoundException();
     }
 
+    if (blog.userId !== ctx.user.id) {
+      throw new ForbiddenException();
+    }
+
     return await this.blogsService.remove(id);
   }
 
-  @Post('blogs/:id/posts')
-  @HttpCode(HttpStatus.CREATED)
   @ApiOperation({ summary: 'Create new post for specific blog' })
   @ApiCreatedResponse({
     status: 201,
@@ -190,14 +199,21 @@ export class BloggerController {
     status: 404,
     description: "If specific blog doesn't exists",
   })
+  @Post('blogs/:id/posts')
+  @HttpCode(HttpStatus.CREATED)
   async createBlogPost(
     @Param('id') blogId: string,
     @Body() { shortDescription, content, title }: CreateBlogPostDto,
+    @GetCurrentJwtContext() ctx: JwtATPayload,
   ) {
     const blog = await this.blogsService.findOne(blogId);
 
     if (!blog) {
       throw new NotFoundException();
+    }
+
+    if (blog.userId !== ctx.user.id) {
+      throw new ForbiddenException();
     }
 
     const post = await this.postsService.create({
@@ -233,16 +249,21 @@ export class BloggerController {
     description: 'Not Found',
   })
   @Put('blogs/:blogId/posts/:postId')
-  @HttpCode(HttpStatus.CREATED)
+  @HttpCode(HttpStatus.NO_CONTENT)
   async updateBlogPost(
     @Param('blogId') blogId: string,
     @Param('postId') postId: string,
     @Body() updatePostDto: UpdatePostDto,
+    @GetCurrentJwtContext() ctx: JwtATPayload,
   ) {
     const blog = await this.blogsService.findOne(blogId);
 
     if (!blog) {
-      throw new BadRequestException();
+      throw new NotFoundException();
+    }
+
+    if (blog.userId !== ctx.user.id) {
+      throw new ForbiddenException();
     }
 
     const post = await this.postsService.update(postId, blogId, updatePostDto);
@@ -272,15 +293,20 @@ export class BloggerController {
     description: 'Not Found',
   })
   @Delete('blogs/:blogId/posts/:postId')
-  @HttpCode(HttpStatus.CREATED)
+  @HttpCode(HttpStatus.NO_CONTENT)
   async deleteBlogPost(
     @Param('blogId') blogId: string,
     @Param('postId') postId: string,
+    @GetCurrentJwtContext() ctx: JwtATPayload,
   ) {
     const blog = await this.blogsService.findOne(blogId);
 
     if (!blog) {
-      throw new BadRequestException();
+      throw new NotFoundException();
+    }
+
+    if (blog.userId !== ctx.user.id) {
+      throw new ForbiddenException();
     }
 
     const isDeleted = await this.postsService.remove(postId);
